fewshotbench/requirements.txt
numpy==1.24.3
pandas==1.5.3
torch==2.0.1
torchvision==0.15.2
wandb==0.15.4
anndata==0.9.1
h5py==3.8.0
matplotlib==3.7.1
scanpy==1.9.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tensorboard==2.13.0
notebook==6.5.4
hydra-core==1.3.2
prettytable==3.8.0
hydra-joblib-launcher==1.2.0
goatools==1.3.1
pyarrow==10.0.1
pyfastx~=0.8.4
einops~=0.6.0
dataclasses-json
dataclasses
torch_geometric
biopython